requirements.txt
flask
flask-smorest
flask-socketio
flask-login
sqlalchemy
alembic
python-dotenv
